requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
requests>=2.31.0
python-dotenv>=1.0.0
openpyxl>=3.1.2
pandas>=2.1.0
python-multipart>=0.0.6
urllib3>=2.0.0
